# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Social App API
  description: API documentation for the Social App backend
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /tasks/{taskId}:
    get:
      tags:
      - Tasks
      summary: Get a task by ID
      description: "Retrieves a task by its ID, including resolved requester and helper\
        \ names"
      operationId: getTaskById
      parameters:
      - name: taskId
        in: path
        description: ID of the task to retrieve
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      responses:
        "200":
          description: Task found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskDetailDTO"
        "404":
          description: Task not found
      security:
      - BearerAuth: []
    put:
      tags:
      - Tasks
      summary: Update task by ID
      description: Updates an existing task by its ID
      operationId: updateTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to update
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
        required: true
      responses:
        "200":
          description: Task updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Task not found
        "400":
          description: Invalid task data
      security:
      - BearerAuth: []
    delete:
      tags:
      - Tasks
      summary: Delete a task by ID
      description: Deletes a task by its ID
      operationId: deleteTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task to delete
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      responses:
        "200":
          description: Task deleted successfully
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Task not found
      security:
      - BearerAuth: []
...
  /api/taskinvoices/{id}:
    get:
      tags:
      - Task Invoices
      summary: Get task invoice by ID
      description: Retrieves a single task invoice by its ID
      operationId: getTaskInvoiceById
      parameters:
      - name: id
        in: path
        description: ID of the task invoice to retrieve
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      responses:
        "200":
          description: Task invoice found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskInvoice"
        "404":
          description: Task invoice not found
      security:
      - BearerAuth: []
    put:
      tags:
      - Task Invoices
      summary: Update a task invoice
      description: Updates an existing task invoice by its ID
      operationId: updateTaskInvoice
      parameters:
      - name: id
        in: path
        description: ID of the task invoice to update
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInvoice"
        required: true
      responses:
        "200":
          description: Task invoice updated successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Task invoice not found
        "400":
          description: Invalid task invoice data
        "401":
          description: Invalid or expired Firebase token
      security:
      - BearerAuth: []
    delete:
      tags:
      - Task Invoices
      summary: Delete a task invoice
      description: Deletes a task invoice by its ID
      operationId: deleteTaskInvoice
      parameters:
      - name: id
        in: path
        description: ID of the task invoice to delete
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      responses:
        "204":
          description: Task invoice deleted successfully
        "404":
          description: Task invoice not found
      security:
      - BearerAuth: []
...
  /api/taskinvoices:
    get:
      tags:
      - Task Invoices
      summary: Get all task invoices
      description: Retrieves a list of all task invoices
      operationId: getAllTaskInvoices
      responses:
        "200":
          description: Successfully retrieved task invoices
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskInvoice"
      security:
      - BearerAuth: []
    post:
      tags:
      - Task Invoices
      summary: Create a new task invoice
      description: Creates a new task invoice for the authenticated user
      operationId: createTaskInvoice
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInvoice"
      responses:
        "201":
          description: Task invoice created successfully
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Invalid task invoice data
        "401":
          description: Invalid or expired Firebase token
      security:
      - BearerAuth: []
...
  /api/chats/task/{taskId}:
    post:
      tags:
      - Chats
      summary: Get or create chat for task
      description: "Gets the existing chat for a task, or creates one if none exists\
        \ yet"
      operationId: getCreateChatForTask
      parameters:
      - name: taskId
        in: path
        description: ID of the task
        required: true
        schema:
          type: integer
          format: int32
        example: 1
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Existing chat returned
          content:
            '*/*':
              schema:
                type: object
        "201":
          description: New chat created
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
        "403":
          description: Caller is not a participant in this task
        "404":
          description: Task not found
        "409":
          description: Task has no assigned helper/dependent yet
...
components:
  schemas:
    Address:
      type: object
      properties:
        addressid:
          type: integer
          format: int32
        streetNumber:
          type: integer
          format: int32
        street:
          type: string
        zipcode:
          type: integer
          format: int32
        neighbourhoodid:
          $ref: "#/components/schemas/Location"
    Admin:
      type: object
      properties:
        adminid:
          type: integer
          format: int32
        admincreatedate:
          type: string
          format: date-time
        adminaccesslevel:
          type: integer
          format: int32
        userid:
          $ref: "#/components/schemas/User"
    Badges:
      type: object
      properties:
        badgeid:
          type: integer
          format: int32
        badgeName:
          type: string
        badge_description:
          type: string
        isSpecialist:
          type: boolean
        xpReward:
          type: integer
          format: int32
        ratingid:
          $ref: "#/components/schemas/Ratings"
        badgeDescription:
          type: string
    Compatibility:
      type: object
      properties:
        compatibilityid:
          type: integer
          format: int32
        helperid:
          $ref: "#/components/schemas/Helper"
        dependentid:
          $ref: "#/components/schemas/Dependent"
        compatibilityScore:
          type: integer
          format: int32
        compatibilityColour:
          type: string
        compadibilityid:
          type: integer
          format: int32
          writeOnly: true
    Dependent:
      type: object
      properties:
        dependentId:
          type: integer
          format: int32
        userId:
          $ref: "#/components/schemas/User"
        taskTypeId:
          $ref: "#/components/schemas/TaskType"
    Helper:
      type: object
      properties:
        helperid:
          type: integer
          format: int32
        userid:
          $ref: "#/components/schemas/User"
        taskTypeid:
          $ref: "#/components/schemas/TaskType"
        badgeid:
          $ref: "#/components/schemas/Badges"
        helperXp:
          type: integer
          format: int32
        available:
          type: boolean
    Location:
      type: object
      properties:
        locationid:
          type: integer
          format: int32
        locationCenterPoint:
          type: integer
          format: int32
        locationRadius:
          type: integer
          format: int32
        neighbourhoodid:
          type: integer
          format: int32
        neighbourhoodName:
          type: string
    Ratings:
      type: object
      properties:
        ratingid:
          type: integer
          format: int32
        ratingReview:
          type: string
        totalXpLevel:
          type: integer
          format: int32
        currentGroup:
          type: string
    Task:
      type: object
      properties:
        taskId:
          type: integer
          format: int32
        helperId:
          type: integer
          format: int32
        dependentId:
          type: integer
          format: int32
        locationId:
          type: integer
          format: int32
        taskTypeId:
          type: integer
          format: int32
        needsSpecialist:
          type: boolean
        signedAdminId:
          type: integer
          format: int32
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        helperBadgeId:
          type: integer
          format: int32
        dependentRatingId:
          type: string
        helperRatingId:
          type: string
        adminReview:
          type: string
        compatibilityId:
          type: integer
          format: int32
        status:
          type: string
        title:
          type: string
        instructions:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/TaskImage"
        immediate:
          type: boolean
    TaskImage:
      type: object
      properties:
        taskImageId:
          type: integer
          format: int32
        taskid:
          $ref: "#/components/schemas/TaskInvoice"
        imageUrl:
          type: string
        uploadedAt:
          type: string
          format: date-time
    TaskInvoice:
      type: object
      properties:
        taskid:
          type: integer
          format: int32
        helperid:
          $ref: "#/components/schemas/Helper"
        dependentid:
          $ref: "#/components/schemas/Dependent"
        locationid:
          $ref: "#/components/schemas/Location"
        tasktypeid:
          $ref: "#/components/schemas/TaskType"
        needsspecialist:
          type: boolean
        signedadminid:
          $ref: "#/components/schemas/Admin"
        startdate:
          type: string
          format: date
        enddate:
          type: string
          format: date
        helperbadgeid:
          $ref: "#/components/schemas/Badges"
        dependentRatingreview:
          type: string
        helperRatingreview:
          type: string
        adminReview:
          type: string
        status:
          type: string
        compatibilityid:
          $ref: "#/components/schemas/Compatibility"
        title:
          type: string
        instructions:
          type: string
        adminreview:
          type: string
          writeOnly: true
        immediate:
          type: boolean
        isImmediate:
          type: boolean
    TaskType:
      type: object
      properties:
        tasktypeid:
          type: integer
          format: int32
        badgeid:
          $ref: "#/components/schemas/Badges"
        description:
          type: string
        needsSpecialist:
          type: boolean
        xpWorth:
          type: integer
          format: int32
    User:
      type: object
      properties:
        userid:
          type: integer
          format: int32
        firebaseUid:
          type: string
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
        addressid:
          $ref: "#/components/schemas/Address"
        badgeid:
          $ref: "#/components/schemas/Badges"
        ratingid:
          $ref: "#/components/schemas/Ratings"
        userType:
          type: string
        isAdmin:
          type: boolean
...
    TaskDetailDTO:
      type: object
      properties:
        taskId:
          type: integer
          format: int32
        helperId:
          type: integer
          format: int32
        dependentId:
          type: integer
          format: int32
        locationId:
          type: integer
          format: int32
        taskTypeId:
          type: integer
          format: int32
        needsSpecialist:
          type: boolean
        signedAdminId:
          type: integer
          format: int32
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        helperBadgeId:
          type: integer
          format: int32
        dependentRatingId:
          type: string
        helperRatingId:
          type: string
        adminReview:
          type: string
        compatibilityId:
          type: integer
          format: int32
        status:
          type: string
        requesterName:
          type: string
        helperName:
          type: string
        requesterUserId:
          type: integer
          format: int32
        title:
          type: string
        instructions:
          type: string
        immediate:
          type: boolean
...